refactor the route and shapes into their own module
